requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
google-cloud-firestore==1.6.1
gunicorn==20.0.4
Werkzeug==0.16.0
